requirements.txt
Django==3.2
django-filter==23.5
djangorestframework==3.14.0
